test: wait for the reap event instead of asserting immediate ESRCH in preset process service tests
close()/timeout resolve on a bounded best-effort kill (SIGKILL plus a
100ms grace fallback), so under load the child can still be an unreaped
zombie when they resolve; the immediate kill(pid, 0) ESRCH assertion
probed that window and flaked on integration-shard 6. Wait for the reap
(bounded poll until ESRCH) instead, matching the kernel settlement
tests' pattern.

Split the inherited-entrypoint test into its own file with a shared
fixtures module to stay under the test-file complexity ceiling.
