+++ packages/preset/test/preset-process-service.test.ts
@@ -1,31 +1,14 @@
 // harness-test-tier: integration
 import assert from "node:assert/strict";
-import {
-  existsSync,
-  mkdirSync,
-  mkdtempSync,
-  readFileSync,
-  rmSync,
-  watch,
-  writeFileSync,
-} from "node:fs";
+import { existsSync, mkdtempSync, readFileSync, rmSync, watch } from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import test from "node:test";
-import { INITIAL_SETTINGS_V1 } from "../../kernel/src/index.ts";
-import {
-  createPresetProcessService as createProjectedPresetProcessService,
-  installPresetPackage,
-  type PresetProcessServiceOptions,
-} from "../src/index.ts";
-
-const createPresetProcessService = (options: Omit<PresetProcessServiceOptions, "readSettings">) =>
-  createProjectedPresetProcessService({ ...options, readSettings: () => INITIAL_SETTINGS_V1 });
+import { installPresetPackage } from "../src/index.ts";
+import { createPresetProcessService, waitFor, write } from "./preset-process-service.fixtures.ts";
 
 test("start durably returns admitted before spawn and status exposes every daemon-owned phase", async () => {
-  const fixture = scriptedPackage(
-      `console.log(JSON.stringify({ schema: "preset-script-result/v1", produces: [] }));`,
-    ),
+  const fixture = scriptedPackage(`console.log(JSON.stringify({ schema: "preset-script-result/v1", produces: [] }));`),
     service = createPresetProcessService({
       rootDir: fixture.rootDir,
       userRoot: fixture.userRoot,
@@ -50,13 +33,7 @@ test("start durably returns admitted before spawn and status exposes every daemo
       ({ outcome }) => outcome === "applied",
     );
     assert.equal(terminal.outcome, "applied");
-    assert.deepEqual(terminal.phases, [
-      "admitted",
-      "spawned",
-      "running",
-      "publishing",
-      "applied",
-    ]);
+    assert.deepEqual(terminal.phases, ["admitted", "spawned", "running", "publishing", "applied"]);
     assert.match(terminal.resultDigest ?? "", /^sha256:[0-9a-f]{64}$/u);
   } finally {
     await service.close();
@@ -64,65 +41,8 @@ test("start durably returns admitted before spawn and status exposes every daemo
   }
 });
 
-test("an inherited entrypoint stages its declaring package and detects post-admission changes", async () => {
-  const fixture = inheritedScriptPackage(),
-    service = createPresetProcessService({
-      rootDir: fixture.rootDir,
-      userRoot: fixture.userRoot,
-      timeoutMs: 1_000,
-      publish: async () => {
-        throw new Error("unexpected produce");
-      },
-    });
-  try {
-    const started = await service.start({
-      presetId: "leaf-canary",
-      entrypoint: "check",
-      inputs: { title: "Inherited" },
-      idempotencyKey: "inherit-ok",
-    });
-    assert.equal(started.phase, "admitted");
-    const applied = await waitFor(
-      "inherited preset run to reach a terminal outcome",
-      () => service.status(started.runId),
-      ({ outcome }) => outcome === "applied" || outcome === "failed",
-    );
-    assert.equal(applied.outcome, "applied", JSON.stringify(applied));
-    assert.deepEqual(applied.phases, [
-      "admitted",
-      "spawned",
-      "running",
-      "publishing",
-      "applied",
-    ]);
-    const admitted = await service.start({
-      presetId: "leaf-canary",
-      entrypoint: "check",
-      inputs: { title: "Tamper" },
-      idempotencyKey: "inherit-tamper",
-    });
-    assert.equal(admitted.phase, "admitted");
-    write(
-      path.join(fixture.parentObject, "scripts/check.mjs"),
-      "process.exit(0);",
-    );
-    const failed = await waitFor(
-      "changed preset package rejection",
-      () => service.status(admitted.runId),
-      ({ outcome }) => outcome === "failed",
-    );
-    assert.equal(failed.code, "package_changed");
-    assert.deepEqual(failed.phases, ["admitted", "failed"]);
-  } finally {
-    await service.close();
-    fixture.cleanup();
-  }
-});
-
 test("admission rejects missing entrypoint, bad input, missing capability, and invalid user shadow within one second", async () => {
-  const valid = scriptedPackage(
-      `console.log(JSON.stringify({ schema: "preset-script-result/v1", produces: [] }));`,
-    ),
+  const valid = scriptedPackage(`console.log(JSON.stringify({ schema: "preset-script-result/v1", produces: [] }));`),
     missing = scriptedPackage("", {
       requires: [{ id: "cap:missing/v1", kind: "command", version: "1" }],
     });
@@ -178,17 +98,11 @@ test("admission rejects missing entrypoint, bad input, missing capability, and i
     assert.equal(performance.now() - started < 1_000, true);
     assert.deepEqual(
       [entrypoint.code, badInput.code, capability.code, shadow.code],
-      [
-        "entrypoint_not_found",
-        "invalid_input",
-        "missing_provider",
-        "shadow_invalid",
-      ],
+      ["entrypoint_not_found", "invalid_input", "missing_provider", "shadow_invalid"],
     );
     assert.equal(
       [entrypoint, badInput, capability, shadow].every(
-        ({ outcome, phases }) =>
-          outcome === "op_rejected" && phases.join() === "op_rejected",
+        ({ outcome, phases }) => outcome === "op_rejected" && phases.join() === "op_rejected",
       ),
       true,
     );
@@ -206,18 +120,10 @@ test("bounded child protocol classifies every failure without a silent phase", a
   const cases = [
     ["never-output", "setInterval(() => {}, 1_000);", "timeout"],
     ["nonzero", "process.exit(7);", "child_exit"],
-    [
-      "signal",
-      'process.kill(process.pid, "SIGTERM");',
-      process.platform === "win32" ? "child_exit" : "child_signal",
-    ],
+    ["signal", 'process.kill(process.pid, "SIGTERM");', process.platform === "win32" ? "child_exit" : "child_signal"],
     ["malformed", 'console.log("{");', "malformed_result"],
     ["oversize", 'process.stdout.write("x".repeat(200));', "result_oversize"],
-    [
-      "disconnect",
-      "process.stdout.end(); setInterval(() => {}, 1_000);",
-      "child_disconnect",
-    ],
+    ["disconnect", "process.stdout.end(); setInterval(() => {}, 1_000);", "child_disconnect"],
   ] as const;
   for (const [name, script, code] of cases) {
     const fixture = scriptedPackage(script),
@@ -273,19 +179,11 @@ test("timeout forcibly reaps a child that ignores SIGTERM before publishing term
         inputs: { title: "Canary" },
         idempotencyKey: "ignore-term-timeout",
       }),
-      staging = path.join(
-        fixture.rootDir,
-        ".harness/preset-runs/staging",
-        started.runId,
-      ),
+      staging = path.join(fixture.rootDir, ".harness/preset-runs/staging", started.runId),
       pidPath = path.join(staging, "child.pid"),
       termPath = path.join(staging, "term.seen"),
       termSeen = process.platform === "win32" ? null : observeFileAppearance(termPath);
-    await waitFor(
-      "child.pid file for timeout fixture",
-      () => existsSync(pidPath),
-      Boolean,
-    );
+    await waitFor("child.pid file for timeout fixture", () => existsSync(pidPath), Boolean);
     pid = Number(readFileSync(pidPath, "utf8"));
     if (termSeen) await termSeen;
     // On Windows kill(SIGTERM) terminates unconditionally, so the handler phase cannot be witnessed;
@@ -298,7 +196,7 @@ test("timeout forcibly reaps a child that ignores SIGTERM before publishing term
       ({ outcome }) => outcome === "failed",
     );
     assert.equal(terminal.code, "timeout");
-    assert.throws(() => process.kill(pid!, 0), { code: "ESRCH" });
+    await assertReaped(pid!);
   } finally {
     if (pid)
       try {
@@ -331,39 +229,25 @@ test("close forcibly reaps a child that ignores SIGTERM within a fixed bound", a
         inputs: { title: "Canary" },
         idempotencyKey: "ignore-term-close",
       }),
-      staging = path.join(
-        fixture.rootDir,
-        ".harness/preset-runs/staging",
-        started.runId,
-      ),
+      staging = path.join(fixture.rootDir, ".harness/preset-runs/staging", started.runId),
       pidPath = path.join(staging, "child.pid"),
       termPath = path.join(staging, "term.seen"),
       termSeen = process.platform === "win32" ? null : observeFileAppearance(termPath);
-    await waitFor(
-      "child.pid file for close fixture",
-      () => existsSync(pidPath),
-      Boolean,
-    );
+    await waitFor("child.pid file for close fixture", () => existsSync(pidPath), Boolean);
     pid = Number(readFileSync(pidPath, "utf8"));
     const before = performance.now(),
       closing = service.close();
-    if (process.platform === "win32")
-      assert.equal(
-        terminalOutcome(service.status(started.runId).outcome),
-        false,
-      ); // Windows terminates unconditionally once close() runs, so only the pre-reap state is observable in flight.
+    if (process.platform === "win32") assert.equal(terminalOutcome(service.status(started.runId).outcome), false);
+    // Windows terminates unconditionally once close() runs, so only the pre-reap state is observable in flight.
     else {
       await termSeen;
       assert.doesNotThrow(() => process.kill(pid!, 0));
-      assert.equal(
-        terminalOutcome(service.status(started.runId).outcome),
-        false,
-      );
+      assert.equal(terminalOutcome(service.status(started.runId).outcome), false);
     }
     await closing;
     assert.equal(performance.now() - before < 1_000, true);
     assert.equal(service.status(started.runId).outcome, "outcome_unknown");
-    assert.throws(() => process.kill(pid!, 0), { code: "ESRCH" });
+    await assertReaped(pid!);
   } finally {
     if (pid)
       try {
@@ -421,10 +305,7 @@ test("idempotency never respawns and restart recovery stays outcome_unknown with
       });
     assert.equal(unknown.outcome, "outcome_unknown");
     assert.equal(replay.outcome, "outcome_unknown");
-    assert.equal(
-      unknown.phases.filter((phase) => phase === "spawned").length,
-      1,
-    );
+    assert.equal(unknown.phases.filter((phase) => phase === "spawned").length, 1);
     assert.equal(unknown.phases.at(-1), "outcome_unknown");
   } finally {
     await recovered?.close();
@@ -475,9 +356,7 @@ test("permission staging denies authored, Git, and raw paths while declared prod
   const producer = scriptedPackage(
       'console.log(JSON.stringify({ schema: "preset-script-result/v1", produces: [{ capabilityId: "policy:task-create/v1", payload: { title: "Produced", taskId: "task-produced" } }] }));',
       {
-        produces: [
-          { id: "policy:task-create/v1", kind: "command", version: "1" },
-        ],
+        produces: [{ id: "policy:task-create/v1", kind: "command", version: "1" }],
       },
     ),
     published: Array<Record<string, unknown>> = [],
@@ -504,19 +383,14 @@ test("permission staging denies authored, Git, and raw paths while declared prod
         ({ outcome }) => outcome === "applied",
       );
     assert.equal(result.outcome, "applied");
-    assert.deepEqual(published, [
-      { kind: "task-create", title: "Produced", taskId: "task-produced" },
-    ]);
+    assert.deepEqual(published, [{ kind: "task-create", title: "Produced", taskId: "task-produced" }]);
   } finally {
     await service.close();
     producer.cleanup();
   }
 });
 
-function scriptedPackage(
-  script: string,
-  entrypoint: Record<string, unknown> = {},
-) {
+function scriptedPackage(script: string, entrypoint: Record<string, unknown> = {}) {
   const rootDir = mkdtempSync(path.join(tmpdir(), "ha-preset-process-")),
     userRoot = path.join(rootDir, ".harness/presets"),
     source = path.join(rootDir, "source/user-canary");
@@ -567,83 +441,25 @@ function scriptedPackage(
     cleanup: () => rmSync(rootDir, { recursive: true, force: true }),
   };
 }
-function inheritedScriptPackage() {
-  const rootDir = mkdtempSync(path.join(tmpdir(), "ha-preset-inherited-")),
-    userRoot = path.join(rootDir, ".harness/presets"),
-    source = path.join(rootDir, "source"),
-    manifest = (id: string, extra: Record<string, unknown>) => ({
-      schema: "preset-manifest/v3",
-      id,
-      title: id,
-      vertical: "software/coding",
-      version: "3.0.0",
-      kind: "process-action",
-      outputShape: "repository-diff",
-      kernelVersionRange: { min: "1.0.0" },
-      capabilityImports: [],
-      profiles: [
-        {
-          id: "baseline",
-          title: "Baseline",
-          completionGates: [],
-          templateSelections: [],
-        },
-      ],
-      defaultProfile: "baseline",
-      ...extra,
-    });
-  const parent = path.join(source, "base-canary"),
-    leaf = path.join(source, "leaf-canary");
-  write(
-    path.join(parent, "preset.json"),
-    JSON.stringify(
-      manifest("base-canary", {
-        entrypoints: {
-          check: {
-            type: "script",
-            intent: "Inherited check",
-            inputs: [{ name: "title", type: "string", required: true }],
-            requires: [],
-            produces: [],
-            sideEffects: [],
-            command: "scripts/check.mjs",
-          },
-        },
-      }),
-    ),
-  );
-  write(
-    path.join(parent, "PRESET.md"),
-    "---\nschema: preset-document/v1\ndescription: Base\nwhenToUse: Run inherited scripts.\n---\n# Base\n",
-  );
-  write(
-    path.join(parent, "scripts/check.mjs"),
-    'console.log(JSON.stringify({ schema: "preset-script-result/v1", produces: [] }));',
-  );
-  write(
-    path.join(leaf, "preset.json"),
-    JSON.stringify(manifest("leaf-canary", { extends: "base-canary" })),
-  );
-  write(
-    path.join(leaf, "PRESET.md"),
-    "---\nschema: preset-document/v1\ndescription: Leaf\nwhenToUse: Inherit the base.\n---\n# Leaf\n",
-  );
-  const installed = installPresetPackage({ source: parent, userRoot });
-  installPresetPackage({ source: leaf, userRoot });
-  return {
-    rootDir,
-    userRoot,
-    parentObject: path.join(userRoot, "preset-objects", installed.digest),
-    cleanup: () => rmSync(rootDir, { recursive: true, force: true }),
-  };
-}
-function write(target: string, body: string): void {
-  mkdirSync(path.dirname(target), { recursive: true });
-  writeFileSync(target, body);
-}
 function terminalOutcome(outcome: string): boolean {
-  return ["applied", "op_rejected", "failed", "outcome_unknown"].includes(
-    outcome,
+  return ["applied", "op_rejected", "failed", "outcome_unknown"].includes(outcome);
+}
+// close()/timeout resolve on a bounded best-effort kill, so the child can still be
+// alive or an unreaped zombie at that moment; ESRCH only becomes observable once
+// the parent's event loop reaps it. Wait for that event instead of assuming the
+// service's grace window always outlasts it.
+async function assertReaped(pid: number): Promise<void> {
+  await waitFor(
+    `child ${pid} to be reaped (kill(pid, 0) reports ESRCH)`,
+    () => {
+      try {
+        process.kill(pid, 0);
+      } catch (error) {
+        return (error as NodeJS.ErrnoException).code === "ESRCH";
+      }
+      return false;
+    },
+    Boolean,
   );
 }
 function observeFileAppearance(target: string): Promise<void> {
@@ -654,9 +470,7 @@ function observeFileAppearance(target: string): Promise<void> {
     const watcher = watch(root, { recursive: true }, (_event, filename) => {
       if (
         !settled &&
-        ((filename !== null &&
-          path.normalize(String(filename)) === relativeTarget) ||
-          existsSync(target))
+        ((filename !== null && path.normalize(String(filename)) === relativeTarget) || existsSync(target))
       ) {
         settled = true;
         watcher.close();
@@ -675,19 +489,4 @@ function observeFileAppearance(target: string): Promise<void> {
       resolve();
     }
   });
-}
-async function waitFor<T>(
-  description: string,
-  read: () => T,
-  done: (value: T) => boolean,
-): Promise<T> {
-  let last: T;
-  for (let attempt = 0; attempt < 500; attempt += 1) {
-    last = read();
-    if (done(last)) return last;
-    await new Promise((resolve) => setTimeout(resolve, 10));
-  }
-  throw new Error(
-    `Timed out waiting for ${description}; last observed: ${JSON.stringify(last!)}`,
-  );
 }
